+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.0.0
 
 * Removes transpilation. Users must now transpile ES6 if needed.
-* Now exports an object instead of a function.
-* Adds `transformUnitless`
+* Adds named `transformUnitless` export
+* Adds named `translate` export
+* Adds named `translateUnitless` export
 * Adds support for `skewX` and `skewY`.
 * Adds default value for `transform` argument.